+++ dwara-db/src/main/java/org/ishafoundation/dwaraapi/db/dao/transactional/ClipDao.java
@@ -1,12 +1,10 @@
 package org.ishafoundation.dwaraapi.db.dao.transactional;
 
-import org.ishafoundation.dwaraapi.db.dao.transactional.custom.JobCustom;
-import org.ishafoundation.dwaraapi.db.model.transactional.Clip;
-import org.ishafoundation.dwaraapi.db.model.transactional.Job;
-import org.springframework.data.repository.CrudRepository;
-
 import java.util.List;
 import java.util.Set;
+
+import org.ishafoundation.dwaraapi.db.model.transactional.Clip;
+import org.springframework.data.repository.CrudRepository;
 
 public interface ClipDao extends CrudRepository<Clip,Integer> {
     List<Clip> findAllByIdIn(Set<Integer> clipId );