+++ dwara-db/src/main/java/org/ishafoundation/dwaraapi/enumreferences/CoreFlow.java
@@ -5,7 +5,8 @@ public enum CoreFlow {
 	core_restore_flow("restore-flow"),
 	core_restore_checksumverify_flow("restore-verify-flow"),
 	core_rewrite_flow("rewrite-flow"),
-	core_bulk_transcode_flow("bulk-transcode-flow");
+	core_bulk_transcode_flow("bulk-transcode-flow"),
+	core_bulk_restore_on_ingest_flow("bulk-restore-on-ingest-flow");
 	
 	private String flowName;