+++ dwara-db/src/main/java/org/ishafoundation/dwaraapi/enumreferences/CoreFlowelement.java
@@ -33,9 +33,11 @@ public enum CoreFlowelement {
 	core_restore_flow_restore("C31", "restore-flow", "restore", null, null, null, 31, true, false, null),
 	core_restore_flow_mkv_mov("C32", "restore-flow", null, "video-digi-2020-mkv-mov-gen", new String[] {"C31"}, null, 32, true, false, "{\"include_if\": {\"artifactclass_regex\": \"video-digi-2020.*\"}}"),
 
-	core_bulk_transcode_flow_restore("C41", "bulk-transcode-flow", "restore", null, null, null, 41, true, false, null),
+	core_bulk_transcode_flow_restore("C41", "bulk-transcode-flow", "restore", null, null, null, 41, true, false, "{\"create_held_jobs\": true}"),
 	core_bulk_transcode_flow_mezzanine_proxy_gen("C42", "bulk-transcode-flow", null, "video-proxy-mezz-gen", new String[] {"C41"}, null, 42, true, false, "{\"pathname_regex\": \"^((?!SUB/).)*$\"}"),
-	core_bulk_transcode_flow_move("C43", "bulk-transcode-flow",null, "restructure-mezz-folder", new String[] {"C42"}, null, 43, true, false, null);	
+	core_bulk_transcode_flow_move("C43", "bulk-transcode-flow",null, "restructure-mezz-folder", new String[] {"C42"}, null, 43, true, false, null),
+	
+	core_bulk_restore_on_ingest_flow_restore("C51", "core_bulk_restore_on_ingest_flow", "restore", null, null, null, 41, true, false, null);
 	
 	private String id;
 	private String flowId;